requirements.txt
Django==2.1.5
Pillow==5.4.1
djangorestframework==3.9.1
markdown==3.0.1
django-filter==2.1.0
django-cors-headers==2.4.0
django-oauth-toolkit==1.2.0